+++ LogicBuilder.EntityFrameworkCore.SqlServer/Repositories/IContextRepository.cs
@@ -1,7 +1,6 @@
 ﻿using LogicBuilder.Data;
 using LogicBuilder.Domain;
 using LogicBuilder.Expressions.Utils.Expansions;
-using Microsoft.EntityFrameworkCore.Query;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -12,11 +11,6 @@ namespace LogicBuilder.EntityFrameworkCore.SqlServer.Repositories
 {
     public interface IContextRepository
     {
-        [Obsolete("Use GetAsync instead.")]
-        Task<ICollection<TModel>> GetItemsAsync<TModel, TData>(Expression<Func<TModel, bool>> filter = null, Expression<Func<IQueryable<TModel>, IQueryable<TModel>>> queryFunc = null, ICollection<Expression<Func<IQueryable<TModel>, IIncludableQueryable<TModel, object>>>> includeProperties = null)
-            where TModel : BaseModel
-            where TData : BaseData;
-
         Task<ICollection<TModel>> GetAsync<TModel, TData>(Expression<Func<TModel, bool>> filter = null, Expression<Func<IQueryable<TModel>, IQueryable<TModel>>> queryFunc = null, SelectExpandDefinition selectExpandDefinition = null)
             where TModel : BaseModel
             where TData : BaseData;
@@ -26,18 +20,6 @@ namespace LogicBuilder.EntityFrameworkCore.SqlServer.Repositories
             where TData : BaseData;
 
         Task<TModelReturn> QueryAsync<TModel, TData, TModelReturn, TDataReturn>(Expression<Func<IQueryable<TModel>, TModelReturn>> queryFunc, SelectExpandDefinition selectExpandDefinition = null)
-            where TModel : BaseModel
-            where TData : BaseData;
-
-        [Obsolete("Use the QueryAsync overload without includeProperties instead.")]
-        Task<TModelReturn> QueryAsync<TModel, TData, TModelReturn, TDataReturn>(Expression<Func<IQueryable<TModel>, TModelReturn>> queryFunc, SelectExpandDefinition selectExpandDefinition,
-            ICollection<Expression<Func<IQueryable<TModel>, IIncludableQueryable<TModel, object>>>> includeProperties = null)
-            where TModel : BaseModel
-            where TData : BaseData;
-
-        [Obsolete("Use the QueryAsync overload without includeProperties instead.")]
-        Task<TReturn> QueryAsync<TModel, TData, TModelReturn, TDataReturn, TReturn>(Expression<Func<IQueryable<TModel>, TModelReturn>> queryFunc, SelectExpandDefinition selectExpandDefinition = null,
-            ICollection<Expression<Func<IQueryable<TModel>, IIncludableQueryable<TModel, object>>>> includeProperties = null)
             where TModel : BaseModel
             where TData : BaseData;